splitgraph.commands removed altogether
