+++ splitgraph/commandline/misc.py
@@ -8,9 +8,8 @@ import splitgraph.core.repository
 from splitgraph import SplitGraphException, CONFIG
 from splitgraph._data.images import _get_all_child_images, delete_images, _get_all_parent_images
 from splitgraph.commandline._common import image_spec_parser
-from splitgraph.commands.repository import repository_exists
 from splitgraph.config.keys import KEYS, SENSITIVE_KEYS
-from splitgraph.core.repository import init_engine, cleanup_objects
+from splitgraph.core.repository import init_engine, cleanup_objects, repository_exists
 from splitgraph.engine import switch_engine
