+++ test/splitgraph/test_commandline.py
@@ -4,10 +4,9 @@ from decimal import Decimal
 
 import pytest
 
-from splitgraph.commands.repository import repository_exists
 from splitgraph.config import PG_PWD, PG_USER
 from splitgraph.core._common import parse_connection_string, serialize_connection_string
-from splitgraph.core.repository import Repository
+from splitgraph.core.repository import Repository, repository_exists
 from splitgraph.engine import switch_engine
 
 try:
